# docs/blueprint/openapi.yaml
openapi: 3.1.0
info:
  title: Approval Workflow Engine API
  version: 1.0.0
  description: API-first contract for the Spring Boot 4 Approval Workflow Engine. Versioning uses the API-Version request header (default 1.0).
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /workflow-definitions:
    parameters:
      - $ref: '#/components/parameters/ApiVersion'
    post:
      tags: [Workflows]
      operationId: createWorkflowDefinition
      summary: Create workflow definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinitionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workflow-definitions/{definitionKey}/versions:
    parameters:
      - $ref: '#/components/parameters/ApiVersion'
    post:
      tags: [Workflows]
      operationId: createWorkflowVersion
      summary: Create workflow draft version
      parameters:
        - $ref: '#/components/parameters/DefinitionKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowVersionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersionResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /workflow-versions/{workflowVersionId}/activate:
    parameters:
      - $ref: '#/components/parameters/ApiVersion'
    post:
      tags: [Workflows]
      operationId: activateWorkflowVersion
      summary: Activate workflow version
      parameters:
        - $ref: '#/components/parameters/WorkflowVersionId'
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersionResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /workflow-versions/{workflowVersionId}:
    parameters:
      - $ref: '#/components/parameters/ApiVersion'
    get:
      tags: [Workflows]
      operationId: getWorkflowVersion
      summary: Get workflow version details
      parameters:
        - $ref: '#/components/parameters/WorkflowVersionId'
      responses:
        '200':
          description: Version details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersionResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApiVersion:
      name: API-Version
      in: header
      required: false
      schema:
        type: string
        default: "1.0"
...
    WorkflowVersionId:
      name: workflowVersionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    DefinitionKey:
      name: definitionKey
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflict or invalid state transition
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    WorkflowVersionStatus:
      type: string
      enum: [DRAFT, ACTIVE, RETIRED]
    ApiError:
      type: object
      required: [code, message, correlationId, details]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
    ApiErrorDetail:
      type: object
      required: [field, reason]
      properties:
        field: { type: string }
        reason: { type: string }
...
    WorkflowDefinitionInput:
      type: object
      required: [definitionKey, name, requestType]
      properties:
        definitionKey: { type: string, maxLength: 100 }
        name: { type: string, maxLength: 150 }
        requestType: { type: string, maxLength: 80 }
        allowLoopback: { type: boolean, default: false }
    WorkflowDefinitionResource:
      type: object
      required: [id, definitionKey, name, requestType, allowLoopback]
      properties:
        id: { type: string, format: uuid }
        definitionKey: { type: string }
        name: { type: string }
        requestType: { type: string }
        allowLoopback: { type: boolean }
    WorkflowVersionInput:
      type: object
      required: [graph]
      properties:
        graph:
          $ref: '#/components/schemas/WorkflowGraph'
    WorkflowVersionResource:
      type: object
      required: [id, definitionKey, versionNo, status, graph]
      properties:
        id: { type: string, format: uuid }
        definitionKey: { type: string }
        versionNo: { type: integer }
        status: { $ref: '#/components/schemas/WorkflowVersionStatus' }
        graph:
          $ref: '#/components/schemas/WorkflowGraph'
        checksumSha256: { type: string, nullable: true }
        activatedAt: { type: string, format: date-time, nullable: true }
    WorkflowGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/WorkflowNode' }
        edges:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/WorkflowEdge' }
        policies:
          type: object
          additionalProperties: true
    WorkflowNode:
      type: object
      required: [id, type]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        type: { $ref: '#/components/schemas/WorkflowNodeType' }
        assignment: { $ref: '#/components/schemas/WorkflowAssignment' }
        ruleRef: { $ref: '#/components/schemas/WorkflowRuleRef' }
        join: { $ref: '#/components/schemas/WorkflowJoin' }
        sla: { $ref: '#/components/schemas/WorkflowSla' }
    WorkflowEdge:
      type: object
      required: [from, to]
      properties:
        from: { type: string, minLength: 1, maxLength: 80 }
        to: { type: string, minLength: 1, maxLength: 80 }
        condition:
          type: object
          additionalProperties: true
    WorkflowAssignment:
      type: object
      required: [strategy]
      properties:
        strategy: { $ref: '#/components/schemas/WorkflowAssignmentStrategy' }
        role: { type: string }
        userId: { type: string }
        expression: { type: string }
    WorkflowRuleRef:
      type: object
      required: [ruleSetKey, version]
      properties:
        ruleSetKey: { type: string, minLength: 1, maxLength: 100 }
        version: { type: integer, minimum: 0 }
    WorkflowJoin:
      type: object
      required: [policy]
      properties:
        policy: { $ref: '#/components/schemas/WorkflowJoinPolicy' }
        quorum: { type: integer, minimum: 1 }
    WorkflowSla:
      type: object
      properties:
        dueInHours: { type: integer, minimum: 1 }
    WorkflowNodeType:
      type: string
      enum: [START, APPROVAL, GATEWAY, JOIN, END]
    WorkflowAssignmentStrategy:
      type: string
      enum: [ROLE, USER, RULE]
    WorkflowJoinPolicy:
      type: string
      enum: [ALL, ANY, QUORUM]
...