# src/main/java/com/isaac/approvalworkflowengine/workflowtemplate/model/WorkflowVersionStatus.java
package com.isaac.approvalworkflowengine.workflowtemplate.model;

public enum WorkflowVersionStatus {
    DRAFT,
    ACTIVE,
    RETIRED;

    public boolean isMutable() {
        return this == DRAFT;
    }
}